Initial commit
Prototype version
Main class serves as a demo for the prototype.

Features added:
Can create Process objects based on process IDs
Can check if Process is alive or not
Can get any file within the Process' /proc directory and return its contents in a List or a Map format
Special object for a Process' stat-files which includes named keys for each value contained in the file.
Special object for an OS's stat-file which includes named keys for each value contained in the file.
Processes can list their Threads, which are also capable of providing information from their /proc files.